requirements.txt
Babel==2.3.4
Flask==0.11.1
Flask-Babel==0.11.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.13.1
Flask-WhooshAlchemy==0.56
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.1.3
Tempita==0.5.2
WTForms==2.1
Werkzeug==0.11.11
Whoosh==2.7.4
blinker==1.4
click==6.6
decorator==4.0.10
itsdangerous==0.24
pbr==1.10.0
psycopg2==2.6.2
python-openid==2.2.5
pytz==2016.7
rauth==0.7.2
requests==2.11.1
six==1.10.0
sqlalchemy-migrate==0.10.0
sqlparse==0.2.2
wsgiref==0.1.2
wtf==0.1
